requirements.txt
lineless_table_rec==0.1.0
opencv_python==4.9.0.80
openpyxl==3.1.5
pandas==2.2.3
PyYAML==6.0.1
rapidocr==2.0.5
table_cls==1.2.0
wired_table_rec==1.2.0
numpy==1.26.4
